refactor(spec): separate page objects in helpers
into workspace, content, authoring.
